% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_major_sites.R
\name{assoc.major.sites}
\alias{assoc.major.sites}
\title{Associating point locations sites with allocated sites}
\usage{
assoc.major.sites(x, y, region_sf)
}
\arguments{
\item{x}{A sf object containing Seaquest sites with original (as recorded)
site coordinates.}

\item{y}{An sf object containing allocated site coordinates.}

\item{region_sf}{An sf object for the region coastal boundary.}
}
\value{
A data.frame with sites_major associated with original sites.
}
\description{
These sites have been manually aggregated where it looked like multiple
designations had been given for the same site.
}
