% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornwall_coast_points.R
\docType{data}
\name{cornwall_coast_points}
\alias{cornwall_coast_points}
\title{Cornwall coastal points}
\format{
An object of class \code{"sf"}
}
\usage{
data(cornwall_coast_points)
}
\description{
A sf object providing points along the coast from projecting reporting rate.
}
\keyword{datasets}
