% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_field_view.R
\name{extract.field.view}
\alias{extract.field.view}
\title{Determine the viewable area from each site and extract enviro data}
\usage{
extract.field.view(
  n,
  sites_major_sf,
  height = 30,
  region_sf,
  region_bf,
  modGrid
)
}
\arguments{
\item{n}{A numeric giving row of sites_major_sf.}

\item{sites_major_sf}{A dataframe giving coordinates of each major site.}

\item{height}{A numeric height above sea level.}

\item{region_sf}{An sf object defining the spatial coastal region of study.}

\item{modGrid}{An sf object defining the grid for analysis created using the
makeGrid function.}
}
\value{
A data.frame.
}
\description{
Determine the viewable area from each site and extract enviro data
}
