% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_predict_reporting_rate.R
\name{gam.predict.rr}
\alias{gam.predict.rr}
\title{Predict reporting rate to distance bands / time periods}
\usage{
gam.predict.rr(m, x, region_sf, position_dist_zone, coarse_res = 5000)
}
\arguments{
\item{m}{A model object produced by mgcv::gam.}

\item{x}{A data.frame with environmental data for predictions.}

\item{coarse_res}{}
}
\value{
A data.frame with spatial predictions.
}
\description{
Predict reporting rate to distance bands / time periods
}
