% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{makeGrid}
\alias{makeGrid}
\title{Make hexagonal grid for national grids}
\usage{
makeGrid(region_sf, cell_size, bufferDist)
}
\arguments{
\item{region_sf}{An sf object defining the spatial region.}

\item{cell_size}{A numeric value specifying the size of the grid cell (m)}

\item{bufferDist}{distance from coast (m)}
}
\value{
An sf object with 'id' used as a unique cell id.
}
\description{
Analysis is conducted on hexagonal grids and this function creates the grid
with a specified area.
}
