% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_species_rr.R
\name{map_species_rr}
\alias{map_species_rr}
\title{Map species reporting rate around the Cornish coast}
\usage{
map_species_rr(x, rsf, focal_species)
}
\arguments{
\item{x}{A data.frame with X, Y, site_major, and fit (from predict).}

\item{rsf}{An sf object defining the geographic region land area.}

\item{focal_species}{A character string giving name of focal species.}
}
\value{
A ggplot object.
}
\description{
Function maps the mean estimated reporting rate (proportion of surveys in
which a species was detected) around the Cornish coast.
}
