% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/south_west_model_grid_1km.R
\docType{data}
\name{modGrid1kmSw}
\alias{modGrid1kmSw}
\title{South West England 1km grid}
\format{
An object of class \code{"sf"}
}
\usage{
data(modGrid1kmSw)
}
\description{
A sf object providing a 1km grid across the ocean surrounding the south west
of England, excluding the land area.
}
\keyword{datasets}
