% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_sst.R
\name{modis_sst}
\alias{modis_sst}
\title{Extract and grid Modis Sea Surface Temperature from MODIS}
\usage{
modis_sst(region_sf, modGrid, yr, mth)
}
\arguments{
\item{region_sf}{An sf object defining the spatial coastal region of study.}

\item{modGrid}{An sf object defining the grid for analysis created using the
makeGrid function.}

\item{yr}{A numeric giving the year of data to extract.}

\item{mth}{A numeric giving the month (1-12) of data to extract.}
}
\value{
A data.frame with monthly sst per grid cell.
}
\description{
Download and extract data from
http://oceandata.sci.gsfc.nasa.gov/opendap/hyrax/MODIST/L3SMI/ to a supplied
grid.
}
