% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sdm_adhoc_data.R
\name{prepare.adhoc.sdm.data}
\alias{prepare.adhoc.sdm.data}
\title{Prepare ad hoc data for modelling}
\usage{
prepare.adhoc.sdm.data(
  x,
  y,
  speciesName,
  tgt_group_spp,
  start_date,
  end_date,
  mths,
  n_bkgrd
)
}
\arguments{
\item{x}{A data.frame containing ad hoc occurrence data (as
recorded) site coordinates}

\item{y}{A data.frame with environmental data gridded.}

\item{speciesName}{Name of focal species.}

\item{tgt_group_spp}{Name of species to be included in the target group. Defaults to all species.}

\item{start_date}{String giving the start date in the format "YYYY/MM/DD".}

\item{end_date}{String giving the end date in the format "YYYY/MM/DD".}

\item{mths}{Numeric vector for the months included in the analysis}

\item{n_bkgrd}{Numeric for the number of background point per year.}
}
\value{
A data.frame.
}
\description{
Prepare ad hoc data for modelling
}
