% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_enviro_view_area.R
\name{prepare.enviro.view.area}
\alias{prepare.enviro.view.area}
\title{Prepared gridded environmental data in viewable area.}
\usage{
prepare.enviro.view.area(x)
}
\arguments{
\item{x}{A sf object of environmental data (i.e. 1km gridded data).}
}
\value{
A data.frame.
}
\description{
Summarise data in viewable area.
}
