% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mod_data.R
\name{prepare.mod.data}
\alias{prepare.mod.data}
\title{Prepare Seaquest data for modelling without distance zones}
\usage{
prepare.mod.data(x, y, a, rsf, speciesName)
}
\arguments{
\item{x}{A data.frame containing Seaquest sites with original (as
recorded) site coordinates}

\item{y}{A data.frame with environmental data gridded.}

\item{a}{A data.frame with allocated sites.}

\item{speciesName}{Name of focal species.}
}
\value{
A data.frame.
}
\description{
Prepare Seaquest data for modelling without distance zones
}
