% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction_zones.R
\name{prepare.pred.zones}
\alias{prepare.pred.zones}
\title{Prepare grid/zones for making predictions from SDM}
\usage{
prepare.pred.zones(modGrid, region_sf, coarse_res = 5000)
}
\arguments{
\item{modGrid}{An sf object defining the grid for analysis created using the
makeGrid function.}

\item{region_sf}{An sf object defining the spatial coastal region of study.}

\item{coarse_res}{A numeric giving the horizontal resolution of coarse grid.}
}
\value{
A data.frame.
modGrid = modGrid1km;  region_sf = sw_bd;   coarse_res = 5000
}
\description{
This function prepares the grids to intersect the environmental data with an
original resolution of 1km and to summarise at a coarser resolution,
suitable for making predictions, i.e. that better match the resolution of
the fitted model. The grids are also clipped at the distance bands
(i.e. 0_1km, 1_2km ...). The environmental variables can then be aggregated
to the coarse grids using dplyr functions post hoc, and weights equal to the
amount of the 1km cell that overlaps the coarser grid can be applied.
}
