% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sdm_sq_data.R
\name{prepare.sdm.sq.data}
\alias{prepare.sdm.sq.data}
\title{Prepare SDM Seaquest data}
\usage{
prepare.sdm.sq.data(x, y, speciesName)
}
\arguments{
\item{x}{A data.frame containing Seaquest sites with original (as
recorded) site coordinates}

\item{y}{A data.frame with environmental data gridded.}

\item{speciesName}{Name of focal species.}
}
\value{
A data.frame.
Fill out the data.frame with each time period a survey took place and
each position_dist_code then calculate the number of surveys per time
period where the target species was detected out of the number of surveys
}
\description{
Prepare SDM Seaquest data
}
