% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_enviro_setup.R
\name{site.enviro.setup}
\alias{site.enviro.setup}
\title{Prepare environmental data for model building}
\usage{
site.enviro.setup(x, y, rsf, grd, h = 30)
}
\arguments{
\item{x}{An sf object of the major site locations}

\item{y}{An sf object of gridded environmental data}

\item{rsf}{An sf object defining the geographic region land area}

\item{grd}{An sf grid object}

\item{h}{A numeric giving the height of observation}

\item{rbf}{An sf object defining the geographic region ocean area}
}
\value{
An sf object
}
\description{
Function crops out viewable areas for each site, including gridded
environmental variables, and returns an
}
