% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocated_sites.R
\docType{data}
\name{sites_allocated}
\alias{sites_allocated}
\title{Allocated common sites}
\format{
An object of class \code{"data.table"}
}
\usage{
data(sites_allocated)
}
\description{
The site information for Seaquest currently contains a lot of inaccurate
information but is being fixed by creating common site names. In the
meantime, sites have been manually aggregated where it looked like multiple
designations had been given for the same site. There common sites are
contained in the package for now, but will be unnessary in the future once
the database has been updated.
}
\keyword{datasets}
