% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_boundary_map.R
\docType{data}
\name{south_west_map}
\alias{south_west_map}
\title{South West England boundary map}
\format{
AN object of class \code{"sf"}
}
\usage{
data(south_west_map)
}
\description{
A sf object showing the land boundary for the south west of England.
}
\examples{
data(south_west_map)
plot(south_west_map)
}
\keyword{datasets}
