% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgpm2grid.R
\name{vgpm2grid}
\alias{vgpm2grid}
\title{Extract and grid VGPM data from Oregon State University}
\usage{
vgpm2grid(region_sf, modGrid, vgpm_path, mth, yr)
}
\arguments{
\item{region_sf}{An sf object defining the spatial coastal region of study.}

\item{modGrid}{An sf object defining the grid for analysis created using the
makeGrid function.}

\item{vgpm_path}{The file path to the vgpm data.}

\item{mth}{A numeric giving the month (1-12) of data to extract.}

\item{yr}{A numeric giving the year of data to extract.}
}
\value{
A data.frame with monthly sst per grid cell.
}
\description{
Extract and grid VGPM data from
http://orca.science.oregonstate.edu/1080.by.2160.monthly.hdf.vgpm.m.chl.m.sst.php
}
