% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_sw_1km.R
\docType{data}
\name{enviro_data_sw_1km}
\alias{enviro_data_sw_1km}
\title{South West England marine environmental data at 1km resolution}
\format{
AN object of class \code{"sf"}
}
\usage{
data(enviro_data_sw_1km)
}
\description{
A data.frame containing sea surface temperature (sst), net primary
productivity (npp), terrain ruggedness index (tri), and depth gridded at
1km resolution.
}
\keyword{datasets}
