% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_predict_reporting_rate.R
\name{gam.predict.rr}
\alias{gam.predict.rr}
\title{Predict reporting rate to distance bands / time periods}
\usage{
gam.predict.rr(
  m,
  x,
  region_sf,
  position_dist_zone = c("0_1km", "1_2km", "2_5km", "5_10km"),
  coarse_res = 5000
)
}
\arguments{
\item{m}{A model object produced by mgcv::gam.}

\item{x}{A data.frame with environmental data for predictions.}

\item{region_sf}{An sf object for the region coastal boundary.}

\item{position_dist_zone}{A vector of character strings giving the distance
bands for predictions.}

\item{coarse_res}{A numeric giving the coarse resolution for predictions.}
}
\value{
A data.frame with spatial predictions.
}
\description{
Predict reporting rate to distance bands / time periods
}
