% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_predict_reporting_rate.R
\name{gam.predict.rr}
\alias{gam.predict.rr}
\title{Predict reporting rate}
\usage{
gam.predict.rr(m, nw)
}
\arguments{
\item{m}{A model object.}

\item{nw}{A data.frame with new data for predictions}
}
\value{
A data.frame.
}
\description{
Predict reporting rate to new data.
}
