% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sdm_sq_data.R
\name{prepare.sdm.sq.data}
\alias{prepare.sdm.sq.data}
\title{Prepare Seaquest data for modelling}
\usage{
prepare.sdm.sq.data(x, y, speciesName)
}
\arguments{
\item{x}{A data.frame containing Seaquest sites with original (as
recorded) site coordinates}

\item{y}{A data.frame with environmental data gridded.}

\item{speciesName}{Name of focal species.}
}
\value{
A data.frame.
}
\description{
Prepare Seaquest data for modelling
}
